refactor(studio): share model display helpers
